reset: intel: Add support for Intel MIPS xrx500 SoC
This adds support for the Intel MIPS interAptiv xrx500 SoCs.
Like the xrx200, the 2nd status register address is shifted, but reset
cell count is 2 as bit offset in reset request and reset status
registers is same.

Also add syscon support, as for xrx200 and xrx500 this registers are
mixed with other functionality.

Signed-off-by: Martin Schiller <ms@dev.tdt.de>